# Conflicts:
#	Presentation/DeviceControl/Components/Common/ItemBase.razor.cs
#	Presentation/DeviceControl/Components/Common/RazorComponentBase.cs
#	Presentation/DeviceControl/Components/Item/ItemActions.razor.cs
#	Presentation/DeviceControl/Components/Item/ItemDates.razor.cs
#	Presentation/DeviceControl/Components/Item/ItemTableBody.razor.cs
#	Presentation/DeviceControl/Components/Item/ItemTableHead.razor.cs
#	Presentation/DeviceControl/Components/Item/ItemTopBar.razor.cs
#	Presentation/DeviceControl/Components/NavMenu.razor.cs
#	Presentation/DeviceControl/Components/Nested/PlusLines/AddPlusLines.razor.cs
#	Presentation/DeviceControl/Components/Nested/PlusLines/PlusLines.razor
#	Presentation/DeviceControl/Components/Nested/PlusLines/PlusLines.razor.cs
#	Presentation/DeviceControl/Components/Nested/PlusNestingFks/PlusNestingFks.razor.cs
#	Presentation/DeviceControl/Components/Section/SectionActions.razor.cs
#	Presentation/DeviceControl/Components/Section/SectionIsMarked.razor.cs
#	Presentation/DeviceControl/Components/Section/SectionTopBar.razor.cs
#	Presentation/DeviceControl/DeviceControl.csproj
#	Presentation/DeviceControl/Features/Layout/MainLayout.razor.cs
#	Presentation/DeviceControl/Locale.cs
#	Presentation/DeviceControl/Pages/Menu/Admins/Access/ItemAccess.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Admins/SqlInfo.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Admins/Versions.razor
#	Presentation/DeviceControl/Pages/Menu/Admins/Versions.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Devices/Hosts/Hosts.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Devices/Lines/ItemLines.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Devices/Lines/Lines.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Logs/MainLogs/ItemLog.razor
#	Presentation/DeviceControl/Pages/Menu/Logs/MainLogs/ItemLog.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Logs/MainLogs/Logs.razor
#	Presentation/DeviceControl/Pages/Menu/Logs/MainLogs/Logs.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Logs/WebServiceLogs/ItemWebServiceLog.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Logs/WebServiceLogs/WebServiceLogs.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Operations/Barcodes/BarCodes.razor
#	Presentation/DeviceControl/Pages/Menu/Operations/Barcodes/BarCodes.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Operations/Barcodes/ItemBarCode.razor
#	Presentation/DeviceControl/Pages/Menu/Operations/Barcodes/ItemBarCode.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Operations/Labels/ItemLabel.razor
#	Presentation/DeviceControl/Pages/Menu/Operations/Labels/Labels.razor
#	Presentation/DeviceControl/Pages/Menu/Operations/PluLabelsAggr.razor
#	Presentation/DeviceControl/Pages/Menu/Operations/PluLabelsAggr.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Operations/PlusLabels/ItemPluLabel.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Operations/PlusLabels/PlusLabels.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Operations/PlusWeightings/ItemPluWeighting.razor
#	Presentation/DeviceControl/Pages/Menu/Operations/PlusWeightings/ItemPluWeighting.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Operations/PlusWeightings/PluWeightings.razor
#	Presentation/DeviceControl/Pages/Menu/Operations/PlusWeightings/PluWeightings.razor.cs
#	Presentation/DeviceControl/Pages/Menu/Profiles/Profile.razor.cs
#	Presentation/DeviceControl/Pages/Menu/References/TemplateResources/TemplateResources.razor.cs
#	Presentation/DeviceControl/Pages/Menu/References/Templates/ItemTemplate.razor.cs
#	Presentation/DeviceControl/Pages/Menu/References/Templates/Templates.razor.cs
#	Presentation/DeviceControl/Pages/Menu/References/Workshops/ItemWorkshop.razor.cs
#	Presentation/DeviceControl/Pages/Menu/References1C/Boxes/Boxes.razor.cs
#	Presentation/DeviceControl/Pages/Menu/References1C/Brands/Brands.razor.cs
#	Presentation/DeviceControl/Pages/Menu/References1C/Bundles/Bundles.razor.cs
#	Presentation/DeviceControl/Pages/Menu/References1C/Clips/Clips.razor.cs
#	Presentation/DeviceControl/Pages/Menu/References1C/Plus/ItemPlu.razor.cs
#	Presentation/DeviceControl/Pages/Menu/References1C/Plus/Plu.razor.cs
#	Presentation/DeviceControl/Pages/_Host.cshtml
#	Presentation/DeviceControl/Program.cs
#	Presentation/DeviceControl/Services/RouteService.cs
#	Presentation/DeviceControl/Utils/BlazorCoreUtils.cs
#	Presentation/DeviceControl/Utils/CssStyleTableHeadModel.cs
#	Presentation/DeviceControl/Utils/RouteUtils.cs
#	Presentation/DeviceControl/_Imports.razor
#	Presentation/ScalesHybrid/MauiProgram.cs

+++ Libs/Ws.Labels/Common/BaseLabelModel.cs
@@ -28,7 +28,7 @@ public class BaseLabelModel : ISerializable
     
     #region Plu
 
-    [XmlElement] public int PluNumber { get; set; }
+    [XmlElement] public short PluNumber { get; set; }
     [XmlElement] public string PluGtin { get; set; } = string.Empty;
     [XmlElement] public string PluFullName { set; get; } = string.Empty;
     [XmlElement] public string PluDescription { get; set; } = string.Empty;
